requirements.txt
requests>=2.31.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
python-telegram-bot>=21.0
APScheduler>=3.10.0
python-dotenv>=1.0.0
tenacity>=8.2.0
pytest>=7.4.0
